started support for Taiko and Ethereum

+++ setup.sh
@@ -115,6 +115,9 @@ getClient() {
 	elif sudo docker ps -a | grep pathfinder > /dev/null
 	then
 		node_docker="pathfinder"
+    elif sudo docker ps -a | grep geth > /dev/null
+	then
+		node_docker="geth"
 	else
 		node_docker="null"
 	fi
@@ -136,7 +139,7 @@ menu_installer() {
             sudo docker logs -f $client &>> $LOGS_PATH & nohup ./myOsiris&
             sleep 2
         fi
-        timeout 20s tail -f nohup.out
+        timeout 10s tail -f nohup.out
     fi
     if [ "${options[$selected]}" = "Stop" ]; then
         osirisClear
@@ -160,10 +163,28 @@ menu_installer() {
 
 menu_running() {
     osirisClear
+    options=("Ethereum" "Starknet" "Taiko" "Quit")
+    yesOrNo=("Yes" "No" "Quit")
+    selected=0 # Initialize the selected variable
+    print_menu "Welcome to myOsiris!" "Please chose a blockchain you'd like to setup" "${options[@]}"
+
+    if [ "${options[$selected]}" = "Starknet" ]; then
+        menu_starknet
+    elif [ "${options[$selected]}" = "Taiko" ]; then
+        installTools
+        installTaiko
+    elif [ "${options[$selected]}" = "Ethereum" ]; then
+        installTools
+        installGeth
+    fi
+}
+
+menu_starknet() {
+    osirisClear
     options=("Papyrus - Starkware" "Pathfinder - Equilibrium" "Juno - Nethermind" "Quit")
     yesOrNo=("Yes" "No" "Quit")
     selected=0 # Initialize the selected variable
-    print_menu "Welcome to myOsiris!" "Please chose the client you'd like to install" "${options[@]}"
+    print_menu "You selected Starknet" "Please chose the client you'd like to install" "${options[@]}"
     # Prompt for node name, rpc_key, and osiris_key
     echo -e -n "${yellow}> Enter a name for your node:${reset} "
     read node_name
@@ -201,10 +222,10 @@ main(){
 
 installPapyrus() {
     osirisClear
-    echo -e "\n\033[34mRunning docker... \033[m"
+    echo -e "\n\033[34mCloning and running docker... \033[m"
     sleep 1
     refreshClient
-    git clone git@github.com:starkware-libs/papyrus.git $CLIENT_DIR &> /dev/null
+    git clone git@github.com:starkware-libs/papyrus.git $CLIENT_DIR
     sudo docker run -d --rm --name papyrus \
         -p 8080-8081:8080-8081 \
         -v $CLIENT_DIR:/app/data \
@@ -229,10 +250,10 @@ installPapyrus() {
 
 installJuno() {
     osirisClear
-    echo -e "\n\033[34mRunning docker... \033[m"
+    echo -e "\n\033[34mCloning and running docker... \033[m"
     sleep 1
     refreshClient
-    git clone https://github.com/NethermindEth/juno $CLIENT_DIR &> /dev/null
+    git clone https://github.com/NethermindEth/juno $CLIENT_DIR
     sudo docker run -d -it --name juno \
         -p 6060:6060 \
         -v $CLIENT_DIR:/var/lib/juno \
@@ -259,10 +280,12 @@ installJuno() {
 
 installPathfinder() {
     osirisClear
-    echo -e "\n\033[34mRunning docker... \033[m"
+    echo -e "\n\033[34mCloning and running docker... \033[m"
     sleep 1
     refreshClient
-    git clone git@github.com:eqlabs/pathfinder.git $CLIENT_DIR &> /dev/null
+    git clone git@github.com:eqlabs/pathfinder.git $CLIENT_DIR
+    sudo mkdir -p $HOME/pathfinder
+    sudo chmod 777 $HOME/pathfinder
     sudo docker run \
         --name pathfinder \
         --restart unless-stopped \
@@ -290,6 +313,94 @@ installPathfinder() {
     fi
 }
 
+installGeth() {
+    osirisClear
+    echo -e "\n\033[34mCloning and running docker... \033[m"
+    sleep 1
+    refreshClient
+    git clone https://github.com/ethereum/go-ethereum.git $CLIENT_DIR
+
+    # Set the data directories for Geth and Nimbus
+    GETH_DATA=$(pwd)/geth_data
+    NIMBUS_DATA=$(pwd)/nimbus_data
+
+    sudo chown -R $(id -u):$(id -g) "$NIMBUS_DATA"
+
+    # Pull the Geth and Nimbus Docker images
+    docker pull ethereum/client-go:latest
+    docker pull statusim/nimbus-eth2:latest
+
+    # Create a custom network for the containers
+    docker network create eth2_network
+
+    # Start the Geth Docker container
+    docker run -d \
+    --name geth \
+    --network eth2_network \
+    -p 30303:30303 \
+    -p 30303:30303/udp \
+    -p 8551:8551 \
+    -u "$(id -u):$(id -g)" \
+    -e RUST_LOG=info \
+    -v "$GETH_DATA:/home/geth/.ethereum" \
+    ethereum/client-go \
+    --datadir /home/geth/.ethereum \
+    --http \
+    --http.addr 0.0.0.0 \
+    --http.port 8551 \
+    --http.vhosts=*
+
+    # Start the Nimbus Docker container and connect it to the Geth container
+    docker run -it \
+    --name nimbus \
+    --network eth2_network \
+    -p 9000:9000 \
+    -v "$NIMBUS_DATA:/home/nimbus/.nimbus" \
+    --user "$(id -u):$(id -g)" \
+    statusim/nimbus-eth2:amd64-latest \
+    --log-level=info \
+    --data-dir=/home/nimbus/.nimbus \
+    --web3-url=http://geth:8551
+
+    # Wait for the Geth client to start
+    osirisClear
+    echo -e "\n\033[34mWaiting for Geth container to be in a running state... \033[m"
+    while [[ "$(sudo docker inspect -f '{{.State.Status}}' geth)" != "running" ]]; do sleep 1; done
+    osirisClear
+    echo -e "\n\033[34mWaiting for Geth client to start... \033[m"
+    while ! sudo docker exec geth grep Ethereum > /dev/null; do sleep 1; done
+    echo "{\"name\": \"${node_name}\", \"client\": \"${client}\", \"rpc_key\": \"${rpc_key}\", \"osiris_key\": \"${osiris_key}\"}" > config.json
+    go build
+    echo -e "\n\033[32m$name full node is running correctly using Geth client!\033[m"
+    if [ $TRACK_MODE == true ]; then
+        sudo docker logs -f geth &>> $LOGS_PATH & nohup ./myOsiris&
+        sleep 2
+        echo -e -n "\n${red}Tracking view mode will exit in 10secs${reset}\n"
+        timeout 10s tail -f nohup.out
+    else
+        exit
+    fi
+}
+
+installTaiko() {
+    osirisClear
+    echo -e "\n\033[34mCloning Taiko node... \033[m"
+    sleep 1
+    refreshClient
+    git clone https://github.com/taikoxyz/simple-taiko-node.git $CLIENT_DIR
+    cd $CLIENT_DIR
+    osirisClear
+    echo -e "\n\033[34mConfiguring Taiko node... \033[m"
+    sleep 1
+    cp .env.sample .env
+    # Prompt the user to set L1_ENDPOINT_HTTP and L1_ENDPOINT_WS in the .env file manually
+    osirisClear
+    echo -e "\n\033[34mStarting Taiko node... \033[m"
+    sleep 1
+    sudo docker compose up -d
+}
+
+
 installTools() {
     osirisClear
     echo -e "\n\033[34mInstalling tools pre-requisites... \033[m\n"
@@ -304,19 +415,26 @@ EOF
 )
     osirisClear
     echo -e "\n\033[34mInstalling tools... \033[m\n"
-    sleep 1
-    sudo apt-get update
-    sudo apt-get install -y \
-        apt-transport-https \
-        ca-certificates \
-        curl \
-        gnupg \
-        lsb-release
-    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
-    echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
-    sudo apt-get update
-    sudo apt-get install -y docker-ce docker-ce-cli containerd.io
-    sleep 1
+    if ! command -v docker &> /dev/null; then
+        sudo apt-get update
+        sudo apt-get install -y \
+            apt-transport-https \
+            ca-certificates \
+            curl \
+            gnupg \
+            lsb-release
+        curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
+        echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list
+        sudo apt-get update
+        sudo apt-get install -y docker-ce docker-ce-cli containerd.io
+    fi
+    if command version go >/dev/null; then
+        echo "Installing go language package version 1.20.2"
+        curl https://storage.googleapis.com/golang/go1.20.2.linux-amd64.tar.gz | sudo tar -C /usr/local -xzf -
+        echo 'export PATH=$PATH:/usr/local/go/bin' | sudo tee -a /etc/environment > /dev/null
+        sudo chmod 0644 /etc/environment
+        source /etc/environment
+    fi
     while read -r p ; do sudo apt install -y $p ; done < <(cat << "EOF"
         sysstat
         bc
@@ -353,6 +471,13 @@ refreshClient()
 	then
 		sudo docker rm -f pathfinder > /dev/null
 		sudo docker image rm -f eqlabs/pathfinder > /dev/null
+        rm -rf ./nohup.out > /dev/null
+		rm -f $LOGS_PATH > /dev/null
+	fi
+    if sudo docker ps -a | grep geth > /dev/null
+	then
+		sudo docker rm -f geth > /dev/null
+		sudo docker image rm -f ethereum/client-go > /dev/null
         rm -rf ./nohup.out > /dev/null
 		rm -f $LOGS_PATH > /dev/null
 	fi
